requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
Deprecated==1.2.13
idna==3.4
Markdown==3.4.1
prompt-toolkit==3.0.37
psutil==5.9.4
pycparser==2.21
PyGithub==1.58.0
PyJWT==2.6.0
PyNaCl==1.5.0
python-dotenv==0.21.1
questionary==1.10.0
requests==2.28.2
soupsieve==2.4
urllib3==1.26.14
wcwidth==0.2.6
wrapt==1.14.1
